WIP: refactor: convert targets into sequential build steps
`build` and `dev` were originally treated as separate jobs because it minimized duplication of code.  However, the jobs could not be run in parallel which meant the build time was proportional to the number of architectures.

Bringing `build` and `dev` into the same job lets all architectures build in parallel.  It also lets `build` and `dev` build from the same local cache.