feat: rewrite release system to support multiple images per commit
A critical limitation of the first design was the assumption that there would only be one Docker build per commit.  As such, our images have become stale over time.

This new design opens the door to better practices.  It allows regular rebuilds of the Docker images to incorporate the latest security fixes and bug patches beyond the base image.