feat: add integration tests
This runs a set of tests before the candidate image can be pushed.  The tests aim to cover all common use cases.

It was not possible to run the tests as a separate set of jobs because `actions/upload-artifact` and `actions/download-artifact` are painfully slow.  Moreover, a separate set of jobs would create a bottleneck because some architectures complete their docker builds significantly quicker than others.  So having the tests inline is probably fastest, albeit not as clean.

PR: #46